Remove netconfetti/ directory, remove whitespace
