--- netconfetti/deliminators.js
@@ -1,5 +1,0 @@
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.beginFrame = /\n#(\d*)\n/;
-exports.endFrame = /(.|\n)+(?=\n##)/;
-exports.endFrameNoCapture = /\n##\n/;
-exports.endHello = /]]>]]>/;
