--- netconfetti/index.js
@@ -1,10 +1,0 @@
-Object.defineProperty(exports, "__esModule", { value: true });
-const client_1 = require("./client");
-
-exports.Client = client_1.default;
-const rpcError_1 = require("./rpcError");
-
-exports.RPCError = rpcError_1.default;
-const rpcErrors_1 = require("./rpcErrors");
-
-exports.RPCErrors = rpcErrors_1.default;
