+++ package.json
@@ -8,6 +8,7 @@
     "lint:fix": "npx eslint --fix ."
   },
   "dependencies": {
+    "@128technology/netconfetti": "^1.0.2",
     "escape-html": "^1.0.3",
     "ssh2": "^1.6.0",
     "uuid": "^8.3.2",
